Moved flow logic from the device to the driver as each device registered the listeners overwriting the old.
